refactor: move component to a different folder
